+++ js/src/jit/IonBuilder.h
@@ -460,7 +460,7 @@ class IonBuilder : public MIRGenerator
                                      ScalarTypeDescr::Type type,
                                      bool canBeNeutered,
                                      MDefinition *value);
-    bool checkTypedObjectIndexInBounds(size_t elemSize,
+    bool checkTypedObjectIndexInBounds(int32_t elemSize,
                                        MDefinition *obj,
                                        MDefinition *index,
                                        TypeDescrSet objTypeDescrs,
@@ -498,7 +498,7 @@ class IonBuilder : public MIRGenerator
                                            TypeDescrSet objTypeReprs,
                                            MDefinition *value,
                                            TypeDescrSet elemTypeReprs,
-                                           size_t elemSize);
+                                           int32_t elemSize);
 
     // jsop_getelem() helpers.
     bool getElemTryDense(bool *emitted, MDefinition *obj, MDefinition *index);
@@ -514,13 +514,13 @@ class IonBuilder : public MIRGenerator
                                            MDefinition *index,
                                            TypeDescrSet objTypeReprs,
                                            TypeDescrSet elemTypeReprs,
-                                           size_t elemSize);
+                                           int32_t elemSize);
     bool getElemTryComplexElemOfTypedObject(bool *emitted,
                                             MDefinition *obj,
                                             MDefinition *index,
                                             TypeDescrSet objTypeReprs,
                                             TypeDescrSet elemTypeReprs,
-                                            size_t elemSize);
+                                            int32_t elemSize);
 
     // Typed array helpers.
     MInstruction *getTypedArrayLength(MDefinition *obj);
